pyproject.toml
[tool.poetry]
name = "ld51-server"
version = "0.1.0"
description = ""
authors = ["Simon Berger <simon@siku2.io>"]
readme = "README.md"
packages = [{ include = "ld51_server" }]

[tool.poetry.dependencies]
fastapi = "^0.85.0"
python = "^3.10"
websockets = "^10.3"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
pylint = "^2.15.3"
pylint-pydantic = "^0.1.4"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
pytest-timeout = "^2.1.0"
requests = "^2.28.1"
uvicorn = "^0.18.3"
pyright = "^1.1.274"

[tool.poe.tasks]
run = "uvicorn ld51_server:app --reload"
test = "pytest"
test-cov = "pytest --cov=ld51_server"
lint = "pylint ld51_server/"
type-check = "pyright"
_sort-imports = "isort ."
_sort-imports-check = "isort --diff --check ."
_format = "black ."
_format-check = "black --check --diff ."
format = ["_sort-imports", "_format"]
format-check = ["_sort-imports-check", "_format-check"]
flow = ["format", "type-check", "test", "lint"]

[tool.isort]
profile = "black"

[tool.pylint.main]
load-plugins = ["pylint_pydantic"]
extension-pkg-whitelis = ["pydantic"]

[tool.pylint.basic]
good-names = ["ws", "x", "y"]

[tool.pylint.message_control]
disable = [
    "line-too-long",
    "missing-docstring",
    "too-few-public-methods",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-locals",
    "too-many-return-statements",
]

[tool.pyright]
strict = ["ld51_server/"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
